requirements.txt
streamlit==1.46.1
pandas==2.3.0
plotly==6.2.0
yfinance==0.2.64
google-genai==1.24.0
textblob==0.19.0
beautifulsoup4==4.13.4
feedparser==6.0.11
trafilatura==2.0.0
streamlit-option-menu==0.4.0
python-dotenv==1.1.1
requests==2.32.4